feat: Implement Vulkan Pipeline Manager and related components
- Added IPipelineManager interface for pipeline management.
- Created VkPipelineManager class for Vulkan-specific pipeline handling.
- Introduced PipelineDesc structure to encapsulate pipeline creation parameters.
- Implemented shader management in VkShaderManager to support SPIR-V compilation.
- Updated RHI interface to include methods for frame graph operations.
- Enhanced SERuntime to initialize Vulkan pipeline and shaders for rendering.
- Added render_core.hpp to define vertex and push constant structures.
- Updated VulkanDevice to manage render pass and command buffer operations.
- Refactored shader compilation to use uint32_t for SPIR-V code representation.
- Added dynamic state management for viewport and scissor in Vulkan rendering.

+++ src/Render/Managers/ShaderManager/Vulkan/VkShaderManager.hpp
@@ -4,21 +4,21 @@
 #include <assert.h>
 
 #include <queue>
-#include <vector>
 #include <shaderc/shaderc.hpp>
+#include <vector>
 
-#include "../IShaderManager.hpp"
 #include "../../../RHI/Vulkan/VulkanConfig.hpp"
+#include "../IShaderManager.hpp"
 
 namespace SE::Render::Shader
 {
 
 struct Shader
 {
-    VkShaderModule       module;
-    std::vector<uint8_t> source;
-    const char*          fileName;
-    SeShaderType         shaderType;
+    VkShaderModule        module;
+    std::vector<uint32_t> source;
+    const char*           fileName;
+    SeShaderType          shaderType;
 };
 
 class VkShaderManager : public IShaderManager
@@ -52,8 +52,8 @@ class VkShaderManager : public IShaderManager
     shaderc::Compiler       m_compiler;
     shaderc::CompileOptions m_options;
 
-    std::vector<uint8_t> compileShader(const char* source, const char* fileName, SeShaderType shaderType);
-    shaderc_shader_kind  toShaderc(SeShaderType type) const
+    std::vector<uint32_t> compileShader(const char* source, const char* fileName, SeShaderType shaderType);
+    shaderc_shader_kind   toShaderc(SeShaderType type) const
     {
         const shaderc_shader_kind shaderTypes[] = {shaderc_vertex_shader, shaderc_fragment_shader};
         return shaderTypes[type];
